requirements.txt
fastapi
uvicorn
sqlalchemy 
PyJWT 
python-dotenv 
pydantic
argon2-cffi